Fix deployed instrumentation dependency tracing
