+++ instrumentation.ts
@@ -1,16 +1,19 @@
-import { defineNodeInstrumentation } from "evlog/next/instrumentation";
-import {
-  getLogEnvironmentContext,
-  getLogServiceName,
-} from "@/shared/lib/logging/config";
+import type { NodeInstrumentationHooks } from "evlog/next/instrumentation";
 
-const environment = getLogEnvironmentContext();
+export async function register() {
+  if (process.env.NEXT_RUNTIME === "nodejs") {
+    // A literal import lets Next.js trace the logger into deployed functions.
+    const instrumentation =
+      await import("@/shared/lib/logging/instrumentationHelpers");
+    await instrumentation.register();
+  }
+}
 
-export const { register, onRequestError } = defineNodeInstrumentation({
-  service: getLogServiceName(),
-  env: environment,
-  pretty: environment.environment !== "production",
-  stringify: true,
-  minLevel: environment.environment === "production" ? "info" : "debug",
-  captureOutput: false,
-});
+export const onRequestError: NodeInstrumentationHooks["onRequestError"] =
+  async (error, request, context) => {
+    if (process.env.NEXT_RUNTIME === "nodejs") {
+      const instrumentation =
+        await import("@/shared/lib/logging/instrumentationHelpers");
+      await instrumentation.onRequestError(error, request, context);
+    }
+  };
